Define get-value for PersistentHashMap
When calling get-value from a REPL using the {} notation for maps,
Clojure seems to create a PersistenArrayMap. When run from tests,
Clojure seems to use a PersistentHashMap. I have a theory the exact type
depends on the size of the map being produced; at any rate, let's cover
all our bases and define a method handler for each.