Run Whitman in a Docker image
Ultimately I'd like to get this running in a Docker image for ease of
use, and so I can deploy it to a host that supporters running
applications as Docker containers. This is a first step towards
"Dockerizing" Whitman.

Right now, Whitman runs under an extremely outdated version of Ubuntu,
14.04, so I am running it in the same Docker container here, but
eventually I hope to upgrade to a newer version of Ubuntu -- or even run
it using Alpine.

Unfortunately, it does not quite work just yet. There is an error
downloading some dependencies, namely data.json, that I will have to
look into.